+++ Release_Notes.md
@@ -75,6 +75,7 @@
     - The modal window is now centered vertically
     - Improved event listener management and resize functionality in the modal window
     - Added resizing to the modal window for the bottom side, right side, and the bottom-right corner
+    - Added support for initial data for template modals
 
 
 ### Known Bugs